requirements.txt
annotated-types==0.7.0
anyio==4.4.0
contourpy==1.2.1
cycler==0.12.1
fastapi==0.112.0
fonttools==4.53.1
idna==3.7
kiwisolver==1.4.5
matplotlib==3.9.1.post1
numpy==2.0.1
packaging==24.1
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.12.2
